requirements.txt
colorama==0.4.6
rich==13.7.0
python-dotenv==1.0.0 